always retransmit even if the connection is good
This should be removed once the server is fixed to send over the lates
connection not a dead connection.  Here is the bug:

 - rpcc sends request A on connection 1
 - rpcs gets request A on connection 1 and blocks
 - rpcc sends some other request, but decides to shutdown connection 1
 - rpcc resends request A on a new connection 2
 - rpcs get request A again on connection 2 but drops it because
still INPROGRESS
 - rpcs unblocks on request A and sends reply on connection 1,
connection 1 is dead and the reply never goes through

At the end both rpcc and rpcs are waiting on each other. It would be
great if when rpcs unblocked, it would have sent the reply on
connection 2 to connection 1.

Thanks to Yandong for catching this.